webinar-demos/live_dashboard_claude_code-2026-04-03/finished_product/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
openai==1.68.2
python-dotenv==1.1.0
watchdog==4.0.0
PyMuPDF==1.26.1
python-docx==1.1.2
